Remove windows preflight checks from setup that are not going to be used in the msi
