+++ pkg/crc/preflight/preflight_checks_common.go
@@ -4,19 +4,14 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"runtime"
 
 	"github.com/code-ready/crc/pkg/crc/adminhelper"
-	"github.com/code-ready/crc/pkg/crc/cache"
 	"github.com/code-ready/crc/pkg/crc/cluster"
 	"github.com/code-ready/crc/pkg/crc/constants"
 	"github.com/code-ready/crc/pkg/crc/logging"
 	"github.com/code-ready/crc/pkg/crc/machine/bundle"
-	"github.com/code-ready/crc/pkg/crc/validation"
-	"github.com/code-ready/crc/pkg/crc/version"
 	"github.com/code-ready/crc/pkg/embed"
 	crcos "github.com/code-ready/crc/pkg/os"
-	"github.com/docker/go-units"
 	"github.com/pkg/errors"
 )
 
@@ -31,44 +26,7 @@ var bundleCheck = Check{
 	labels: None,
 }
 
-var genericPreflightChecks = []Check{
-	{
-		configKeySuffix:  "check-admin-helper-cached",
-		checkDescription: "Checking if crc-admin-helper executable is cached",
-		check:            checkAdminHelperExecutableCached,
-		fixDescription:   "Caching crc-admin-helper executable",
-		fix:              fixAdminHelperExecutableCached,
-
-		labels: None,
-	},
-	{
-		configKeySuffix:  "check-obsolete-admin-helper",
-		checkDescription: "Checking for obsolete admin-helper executable",
-		check:            checkOldAdminHelperExecutableCached,
-		fixDescription:   "Removing obsolete admin-helper executable",
-		fix:              fixOldAdminHelperExecutableCached,
-	},
-	{
-
-		configKeySuffix:  "check-supported-cpu-arch",
-		checkDescription: "Checking if running on a supported CPU architecture",
-		check:            checkSupportedCPUArch,
-		fixDescription:   "CodeReady Containers is only supported on x86_64 hardware",
-		flags:            NoFix,
-
-		labels: None,
-	},
-	{
-		configKeySuffix:  "check-ram",
-		checkDescription: "Checking minimum RAM requirements",
-		check: func() error {
-			return validation.ValidateEnoughMemory(constants.DefaultMemory)
-		},
-		fixDescription: fmt.Sprintf("crc requires at least %s to run", units.HumanSize(float64(constants.DefaultMemory*1024*1024))),
-		flags:          NoFix,
-
-		labels: None,
-	},
+var genericCleanupChecks = []Check{
 	{
 		cleanupDescription: "Removing CRC Machine Instance directory",
 		cleanup:            removeCRCMachinesDir,
@@ -90,22 +48,13 @@ var genericPreflightChecks = []Check{
 
 		labels: None,
 	},
-}
+	{
+		cleanupDescription: "Removing hosts file records added by CRC",
+		cleanup:            removeHostsFileEntry,
+		flags:              CleanUpOnly,
 
-var cleanUpHostsFile = Check{
-	cleanupDescription: "Removing hosts file records added by CRC",
-	cleanup:            removeHostsFileEntry,
-	flags:              CleanUpOnly,
-
-	labels: None,
-}
-
-func checkSupportedCPUArch() error {
-	if runtime.GOARCH != "amd64" {
-		logging.Debugf("GOARCH is %s", runtime.GOARCH)
-		return fmt.Errorf("CodeReady Containers can only run on x86_64 CPUs")
-	}
-	return nil
+		labels: None,
+	},
 }
 
 func checkBundleExtracted() error {
@@ -157,68 +106,12 @@ func fixBundleExtracted() error {
 	return nil
 }
 
-// Check if helper executable is cached or not
-func checkAdminHelperExecutableCached() error {
-	if version.IsMacosInstallPathSet() || version.IsMsiBuild() {
+func removeHostsFileEntry() error {
+	err := adminhelper.CleanHostsFile()
+	if errors.Is(err, os.ErrNotExist) {
 		return nil
 	}
-
-	helper := cache.NewAdminHelperCache()
-	if !helper.IsCached() {
-		return errors.New("crc-admin-helper executable is not cached")
-	}
-	if err := helper.CheckVersion(); err != nil {
-		return errors.Wrap(err, "unexpected version of the crc-admin-helper executable")
-	}
-	logging.Debug("crc-admin-helper executable already cached")
-	return checkSuid(helper.GetExecutablePath())
-}
-
-func fixAdminHelperExecutableCached() error {
-	if version.IsMacosInstallPathSet() || version.IsMsiBuild() {
-		return nil
-	}
-
-	helper := cache.NewAdminHelperCache()
-	if err := helper.EnsureIsCached(); err != nil {
-		return errors.Wrap(err, "Unable to download crc-admin-helper executable")
-	}
-	logging.Debug("crc-admin-helper executable cached")
-	return setSuid(helper.GetExecutablePath())
-}
-
-var oldAdminHelpers = []string{"admin-helper-linux", "admin-helper-darwin", "admin-helper-windows.exe"}
-
-/* These 2 checks can be removed after a few releases */
-func checkOldAdminHelperExecutableCached() error {
-	logging.Debugf("Checking if an older admin-helper executable is installed")
-	for _, oldExecutable := range oldAdminHelpers {
-		oldPath := filepath.Join(constants.CrcBinDir, oldExecutable)
-		if _, err := os.Stat(oldPath); !os.IsNotExist(err) {
-			return fmt.Errorf("Found old admin-helper executable '%s'", oldExecutable)
-		}
-	}
-
-	logging.Debugf("No older admin-helper executable found")
-
-	return nil
-}
-
-func fixOldAdminHelperExecutableCached() error {
-	logging.Debugf("Removing older admin-helper executable")
-	for _, oldExecutable := range oldAdminHelpers {
-		oldPath := filepath.Join(constants.CrcBinDir, oldExecutable)
-		if err := os.Remove(oldPath); err != nil {
-			if !os.IsNotExist(err) {
-				logging.Debugf("Failed to remove  %s: %v", oldPath, err)
-				return err
-			}
-		} else {
-			logging.Debugf("Successfully removed %s", oldPath)
-		}
-	}
-
-	return nil
+	return err
 }
 
 func removeCRCMachinesDir() error {
@@ -241,12 +134,4 @@ func removeOldLogs() error {
 		}
 	}
 	return nil
-}
-
-func removeHostsFileEntry() error {
-	err := adminhelper.CleanHostsFile()
-	if errors.Is(err, os.ErrNotExist) {
-		return nil
-	}
-	return err
 }
