+++ pkg/crc/preflight/preflight_windows_test.go
@@ -11,13 +11,13 @@ import (
 func TestCountConfigurationOptions(t *testing.T) {
 	cfg := config.New(config.NewEmptyInMemoryStorage())
 	RegisterSettings(cfg)
-	assert.Len(t, cfg.AllConfigs(), 15)
+	assert.Len(t, cfg.AllConfigs(), 10)
 }
 
 func TestCountPreflights(t *testing.T) {
-	assert.Len(t, getPreflightChecks(false, false, network.SystemNetworkingMode), 20)
-	assert.Len(t, getPreflightChecks(true, true, network.SystemNetworkingMode), 20)
+	assert.Len(t, getPreflightChecks(false, false, network.SystemNetworkingMode), 15)
+	assert.Len(t, getPreflightChecks(true, true, network.SystemNetworkingMode), 15)
 
-	assert.Len(t, getPreflightChecks(false, false, network.UserNetworkingMode), 21)
-	assert.Len(t, getPreflightChecks(true, true, network.UserNetworkingMode), 21)
+	assert.Len(t, getPreflightChecks(false, false, network.UserNetworkingMode), 16)
+	assert.Len(t, getPreflightChecks(true, true, network.UserNetworkingMode), 16)
 }
